git-svn-id: https://svn.ais.uni-bonn.de/svn/repos/cuv/trunk@4077 9a9a52a9-7681-4bea-9213-143ca318e872

+++ src/vector_ops/vector_ops.cu
@@ -144,7 +144,7 @@ struct tf_dtanh{  __device__  __host__      T operator()(const T& x, const T& a,
 
 // rectifying transferfunctions a is param beta
 template<class T, class A>
-struct tf_rect{  __device__  __host__       T operator()(const T& x, const A& a)      const{ return log(1+ exp(a*x))/a; } };
+struct tf_rect{  __device__  __host__       T operator()(const T& x, const A& a)      const{ return log(1+ expf(a*x))/a; } };
 template<class T, class A>
 struct tf_drect{  __device__  __host__      T operator()(const T& x, const A& a)      const{ return 1-1/(x*expf(a)); } };
